(fix) Bypass Craft for filename overwrite
Turns out this is the most effective way to 'force' our filename
convention. Changing the filename of the assets triggers a cascades of
barrier and code to handle a change of filename, but it this case, we
just need to overwrite it.

We also need to check for the Volume's FS, not the transform one.

This api will change has the function as no reason this live there.