(feat) Add directory operations: exists + delete (#9)
This commits adds a best effort support of directory existance.
It uses the data in the DB to check if the folder exists.

As for deletion of a directory, since there are no directory structure
in Cloudflare Images, we now have a protection against deletion of non
empty folders. They need to be emptied first before the user can delete
them.